face recognition/requirements.txt
tensorflow
tensorflow-hub
Pillow
